Porting subtask of proxy.php to python
proxy.php always extract all project archives and served them for
download. However, this was I/O intense and PHP claims the file
does not exist directly after writing it (even with fflush).

Determining MD5 hashes from temporary files would be a possible
fix; this is not possible in PHP.

In python ``unpack.py`` is a command-line tool to extract project
files.